Basic (very basic) vfs mounts and opens
Currently, the vfs only works with the root mount, and can open existing
files by walking from the root mount point until it finds the file its
looking for. Eventually, it should grab the current processes CWD and
use that as its root if it the path is not rooted.

Eventually, I need to be able to mount a new filesystem into
directories, and walk between them as if it was a single filesystem, but
that is a problem for future me