feat: pthread, fstat, sched deadlock fix
This commit contains a bit too much as usual, but whatever :). I
implemented some basic thread control (creation only) via pctl and a
pthread interface

I also finally added fstat syscall bindings.

The scheduler also had a deadlock where if a process adds a task, it
takes the spinlock on the sched's mlfq. If the process is then
preempted, the scheduler deadlocks with the process that was just
running. I "fixed" this by disabling interrupts within that critical
section. There's still a tiny area where it could be preempted so a
smarter method using some external atomic queue would be smarter.