[riscv] M -> S interrupt forwarding (timer :D)
okay so get this. I was chasing a bug that caused the data pointed
to by mscratch to be corrupted after the first irq or something. Turns
out I was intentionally wiping .bss (where it was stored) after
initializing it... so of course it was being corrupted.

Either way, we are now in supervisor mode, getting interrupts (at least
from the CLINT so far) so maybe a scheduler binding is in order?