cleanup: remove global cpus array
CPUs are now allocated on the core's stack, and are added
to a global linked list. Unfortunately, this means the CPUs
are more than likely out of order, but they can be sorted.

This addition fixes a few bugs with RISC-V's tendancy to
choose the boot CPU with a lottery. (bootcpu != 0, but goes
into cpus[0] anyways)

This will also help significantly when the kernel can boot
on some of the 8 socket xeons I have access to :^)